Bugfix : RoomView and Reservation DAO

+++ src/business/ReservationManager.java
@@ -7,28 +7,35 @@ import java.util.List;
 
 public class ReservationManager {
     private final ReservationDAO reservationDAO;
+    private final RoomManager roomManager;
 
     public ReservationManager() {
         this.reservationDAO = new ReservationDAO();
-    }
-
-    public void addReservation(Reservation reservation) {
-        reservationDAO.addReservation(reservation);
-    }
-
-    public void updateReservation(Reservation reservation) {
-        reservationDAO.updateReservation(reservation);
-    }
-
-    public Reservation getReservationById(int id) {
-        return reservationDAO.getReservationById(id);
+        this.roomManager = new RoomManager();
     }
 
     public List<Reservation> getAllReservationsWithHotelName() {
         return reservationDAO.getAllReservationsWithHotelName();
     }
 
+    public Reservation getReservationById(int id) {
+        return reservationDAO.getReservationById(id);
+    }
+
+    public void addReservation(Reservation reservation) {
+        reservationDAO.addReservation(reservation);
+        roomManager.decreaseRoomStock(reservation.getRoomId(), 1);
+    }
+
+    public void updateReservation(Reservation reservation) {
+        reservationDAO.updateReservation(reservation);
+    }
+
     public void deleteReservation(int id) {
-        reservationDAO.deleteReservation(id);
+        Reservation reservation = reservationDAO.getReservationById(id);
+        if (reservation != null) {
+            reservationDAO.deleteReservation(id);
+            roomManager.increaseRoomStock(reservation.getRoomId(), 1);
+        }
     }
 }
